git-svn-id: file:///Users/pica/Downloads/GitClone/svnrep/suN/HiRep@526 37a69121-b030-0410-a54a-fdca83ffee47

+++ Utils/suN_utils.c
@@ -20,19 +20,19 @@
 
 static void normalize(suNg_vector *v)
 {
-   double fact;
-	 _vector_prod_re_g(fact,*v,*v);
-   fact=1.0f/sqrt(fact);
-   _vector_mul_g(*v, fact, *v);
+  double fact;
+  _vector_prod_re_g(fact,*v,*v);
+  fact=1.0f/sqrt(fact);
+  _vector_mul_g(*v, fact, *v);
 }
 
 
 static void normalize_flt(suNg_vector_flt *v)
 {
-   double fact;
-	 _vector_prod_re_g(fact,*v,*v);
-   fact=1.0/sqrt(fact);
-   _vector_mul_g(*v, fact, *v);
+  float fact;
+  _vector_prod_re_g(fact,*v,*v);
+  fact=1.0/sqrt(fact);
+  _vector_mul_g(*v, fact, *v);
 }
